light: eliminate pytest_plugins
We tried to unify fixture declaration with pytest_plugins, however
it do not fit in our current architecture. Our current approach is
to declare every fixture in ./python_functional/conftest.py as a
global repository.
